feat: Enhance previous consultation display with detailed information, refine consultation management logic, and improve appointment time filtering.

+++ database/migrations/2026_02_05_152542_create_patients_table.php
@@ -11,46 +11,48 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('patients', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('patients')) {
+            Schema::create('patients', function (Blueprint $table) {
+                $table->id();
 
-            //Añadimos campos
-            $table->foreignId('user_id')
-                ->constrained('users')
-                //Si borro un usuario que se borre el paciente también
-                ->onDelete('cascade');
+                //Añadimos campos
+                $table->foreignId('user_id')
+                    ->constrained('users')
+                    //Si borro un usuario que se borre el paciente también
+                    ->onDelete('cascade');
 
-            $table->foreignId('blood_type_id')
-                ->nullable()
-                ->constrained('blood_types')
-                ->onDelete('set null');
+                $table->foreignId('blood_type_id')
+                    ->nullable()
+                    ->constrained('blood_types')
+                    ->onDelete('set null');
 
-            $table->string('allergies')
-                ->nullable();
+                $table->string('allergies')
+                    ->nullable();
 
-            $table->string('chronic_conditions')
-                ->nullable();
+                $table->string('chronic_conditions')
+                    ->nullable();
 
-            $table->string('surgical_history')
-                ->nullable();
+                $table->string('surgical_history')
+                    ->nullable();
 
-            $table->string('family_history')
-                ->nullable();
+                $table->string('family_history')
+                    ->nullable();
 
-            $table->string('observations')
-                ->nullable();
+                $table->string('observations')
+                    ->nullable();
 
-            $table->string('emergency_contact_name')
-                ->nullable();
+                $table->string('emergency_contact_name')
+                    ->nullable();
 
-            $table->string('emergency_contact_phone')
-                ->nullable();
+                $table->string('emergency_contact_phone')
+                    ->nullable();
 
-            $table->string('emergency_contact_relationship')
-                ->nullable();
+                $table->string('emergency_contact_relationship')
+                    ->nullable();
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
