+++ database/migrations/2026_02_21_184207_create_doctors_table.php
@@ -11,26 +11,28 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('doctors', function (Blueprint $table) {
-            $table->id();
-            
-            $table->foreignId('user_id')
-                ->constrained('users')
-                ->onDelete('cascade');
-            
-            $table->foreignId('speciality_id')
-                ->nullable()
-                ->constrained()
-                ->onDelete('set null');
+        if (!Schema::hasTable('doctors')) {
+            Schema::create('doctors', function (Blueprint $table) {
+                $table->id();
+                
+                $table->foreignId('user_id')
+                    ->constrained('users')
+                    ->onDelete('cascade');
+                
+                $table->foreignId('speciality_id')
+                    ->nullable()
+                    ->constrained()
+                    ->onDelete('set null');
 
-            $table->string('medical_license_number')
-                ->nullable();
-            
-            $table->text('biography')
-                ->nullable();
-            
-            $table->timestamps();
-        });
+                $table->string('medical_license_number')
+                    ->nullable();
+                
+                $table->text('biography')
+                    ->nullable();
+                
+                $table->timestamps();
+            });
+        }
     }
 
     /**
